% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(
  data,
  r,
  E,
  Eerror,
  nSpl,
  nbRef,
  Refposcol,
  nCTL,
  CF,
  CalPos,
  trace = FALSE,
  geo = FALSE,
  na.rm = na.rm
)
}
\description{
Customized version of the nrmData function, it is meant to be
 used within the calibration_factors function
}
