% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Function to calculate the amplification efficiency}
\usage{
slope(data, q, r, na.rm = FALSE)
}
\arguments{
\item{data}{data.frame containing row datas (genes in columns, samples in
rows, Cq values).}

\item{q}{numeric, cDNA dilution values.}

\item{r}{numeric, number of qPCR replicates.}

\item{na.rm}{logical, indicating whether NA values should be stripped before
the computation proceeds.}
}
\value{
Efficiency Primer amplification efficiency.

 Slope Slope of the dilution curve.

 Intercept Intercept of the dilution curve
}
\description{
Original EasyqpcR function.
This function calculates the amplification efficiency from
classical qPCR dilution experiment using the Cq values. The Cq values are
plotted against the logarithmized concentration (or dilution) values, a
linear regression line is fit and the efficiency calculated by
E = 10^-1/slope.
}
\examples{

data(Efficiency_calculation)

slope(data = Efficiency_calculation, q=c(1000, 100 ,10, 1, 0.1), r=3,
 na.rm=TRUE)

}
\references{
Jan Hellemans, Geert Mortier, Anne De Paepe, Frank Speleman and
Jo Vandesompele. qBase relative quantification framework and software for
management and automated analysis of real-time quantitative PCR data. Genome
Biology 2007, 8:R19 (doi:10.1186/gb-2007-8-2-r19).
}
\seealso{
You can also see the qpcR package with the calib and calib2
functions.
}
\author{
Sylvain Le pape <sylvain.le.pape@univ-poitiers.fr> (original function author)

Daniel S. Rodríguez-Leon (author of easyqpcr2)
}
