% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totData.R
\name{totData}
\alias{totData}
\title{Aggregation of qPCR biological replicates and data transformation}
\usage{
totData(data, r, geo = TRUE, logarithm = TRUE, base, transformation = TRUE,
 nSpl, linear = TRUE, na.rm = na.rm)
}
\arguments{
\item{data}{data.frame containing row datas (genes in columns, samples
in rows, Cq values).}

\item{r}{numeric, number of qPCR replicates.}

\item{geo}{logical, the function will use the geometrical mean of your
biological replicates if TRUE or the arithmetic mean if FALSE.}

\item{logarithm}{logical, the NRQs will be log-transformed.}

\item{base}{numeric, the logarithmic base (2 or 10).}

\item{transformation}{logical, if TRUE, the transformation procedure for
highly variable biological replicates (but with the same tendency) will be
done.}

\item{nSpl}{numeric, the number of samples.}

\item{linear}{logical, after the transformation procedure done, your raw
data will be normalized (anti-log-transformed).}

\item{na.rm}{logical, indicating whether NA values should be stripped before
the computation proceeds.}
}
\value{
Mean of your qPCR runs.
The geometric (if TRUE for geo) or arithmetic mean of your biological
replicates.

Standard deviations of your qPCR runs.
The standard deviation of your biological replicates.

Standard errors of your qPCR runs.
The standard error of your biological replicates.

Transformed data.
If TRUE for transformation, your raw data will be transformed by the algorithm
of Willems et al. (2008).

Reordered transformed data.
The transformed data reordered by rowname.
}
\description{
Original EasyqpcR function.
This function aggregates qPCR biological replicates and calculates the main
parameters : mean (arithmetic or geometric), the standard deviation and the
standard error from your biological replicates of your experience.
This function has an algorithm published by Willems et al. (2008) which
performs a standardization procedure that can be applied to data sets that
display high variation between biological replicates. This enables proper
statistical analysis and drawing relevant conclusions. The procedure is not
new, it has been used in microarray data analysis and is based on a series
of sequential corrections, including log transformation, mean centering,
and autoscaling.
}
\details{
The standardization procedure used in this function (if TRUE for
the transformation argument) is based on the article of Willems et al.
(2008). This function perform successively thEerror operations :
log-transformation of your raw data, mean of your log-transformed data for
each biological replicate, mean centering for each biological replicate,
standard deviation of each mean-centered biological replicate, autoscaling
of your data, i.e., your mean-centered data for each biological replicate
will be divided by the standard deviation of the mean-centered biological
replicate and then multiplicated by the mean of the standard deviation of
all the biological replicates.
For more information for the way to use this
function, please see the vignette
}
\references{
Jan Hellemans, Geert Mortier, Anne De Paepe, Frank Speleman and
Jo Vandesompele. qBase relative quantification framework and software for
management and automated analysis of real-time quantitative PCR data. Genome
Biology 2007, 8:R19 (doi:10.1186/gb-2007-8-2-r19).
}
\author{
Sylvain Le pape <sylvain.le.pape@univ-poitiers.fr> (original function author)

Daniel S. Rodríguez-Leon (author of easyqpcr2)
}
