% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_efficiency.R
\name{amp_efficiency}
\alias{amp_efficiency}
\title{amp_efficiency}
\usage{
amp_efficiency(data, q, r, na.rm = FALSE)
}
\arguments{
\item{na.rm}{}
}
\value{
list
}
\description{
This is a custom version of the "slope" function of the EasyqpcR package.
It has been adapted to run on recent versions of R, and to calculate the
standard error of the efficiency (not only the efficiency)
}
