% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_factors.R
\name{calibration_factors}
\alias{calibration_factors}
\title{calibration_factors}
\usage{
calibration_factors(
  run.data.df,
  n_replicates,
  reference_genes,
  cals_identifier = "Cal",
  amp_efficiencies,
  CF = NA,
  n_control = NA,
  trace = F,
  geo = T,
  na.rm = T
)
}
\arguments{
\item{na.rm}{}
}
\value{
list
}
\description{
Function to automatically get the calibration
factors of each run
}
