% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnrqs.R
\name{cnrqs}
\alias{cnrqs}
\title{cnrqs}
\usage{
cnrqs(nrqs.list, run.data.df)
}
\arguments{
\item{run.data.df}{}
}
\value{
list
}
\description{
Function to obtain the CNRQs values for the whole qpcr
experiment (batch / run series)
}
