% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Determination of the NF, RQ, NRQ, NRQ scaled to control and their
SE and SD}
\usage{
normalize_data(
  data,
  r,
  E,
  Eerror,
  nSpl,
  nbRef,
  Refposcol,
  nCTL,
  CF,
  CalPos,
  trace = FALSE,
  geo = FALSE,
  na.rm = na.rm
)
}
\arguments{
\item{data}{data.frame containing qPCR Cq values data for each gene per sample per run
per well}

\item{r}{numeric, number of qPCR replicates}

\item{E}{numeric, amplification efficiency values for each gene (follow the same order
of the genes).}

\item{Eerror}{numeric, standard errors of amplification efficiencies for each gene (follow
the same order of the genes).}

\item{nSpl}{numeric, number of samples to analyzed.}

\item{nbRef}{numeric, number of reference genes used.}

\item{Refposcol}{column position of your reference gene(s).}

\item{nCTL}{numeric, number of samples forming your control group.}

\item{CF}{numeric (or object if you have used the calData function from this package),
values of the calibration factors for each gene (follow the same order of
the genes).}

\item{CalPos}{numeric, sample number of your calibrator(s).}

\item{trace}{logical, print additional information.}

\item{geo}{logical, to scale to your control group, the function will use the
geometrical mean if TRUE or the arithmetic mean if FALSE.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the
computation proceeds.}
}
\value{
A list with the following entries:

NRQs normalized to control
 Holds the normalized relative quantities scaled to the control group.

NRQs Holds the normalized relative quantities.

NRQs of your calibrator for this run
 Holds the normalized relative quantities of the calibrator(s).
}
\description{
Customized version of the nrmData function, it is meant to be
 used within the calibration_factors or nrqs functions.

 This function is basically identical to the original nmrData.
 Obviously the name is different, which is in snake case format and is
 more informative regarding what the function does.
 But, unlike nmrData it does not rely on the pakcage plyr being loaded, which
 causes conflicts with the dplyr package. Instead, plyr functions are called
 using '::'.
}
\references{
Jan Hellemans, Geert Mortier, Anne De Paepe, Frank Speleman and
Jo Vandesompele. qBase relative quantification framework and software for
management and automated analysis of real-time quantitative PCR data. Genome
Biology 2007, 8:R19 (doi:10.1186/gb-2007-8-2-r19).
}
\author{
Sylvain Le pape <sylvain.le.pape@univ-poitiers.fr> (original function author)

Daniel S. Rodríguez-Leon (author of easyqpcr2)
}
