% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrqs.R
\name{nrqs}
\alias{nrqs}
\title{nrqs}
\usage{
nrqs(
  run.data.df,
  n_replicates,
  reference_genes,
  cals_identifier = "Cal",
  amp_efficiencies,
  cal.factors.list,
  n_control = NA,
  trace = F,
  geo = T,
  na.rm = T
)
}
\arguments{
\item{na.rm}{}
}
\description{
Function to automatically calculate the NRQs values for each run
}
